\name{ddionrails-testsuite-package}
\alias{ddionrails-testsuite-package}
\alias{ddionrails-testsuite}
\docType{package}
\title{
Testing DDI on Rails' CSV-Format Data Import
}
\description{
Several tests to test the correctness of the CSV tables for DDI on Rails' data import. In case of errors those are listed and written into new CSV files.
}
\details{
\tabular{ll}{
Package: \tab ddionrails-testsuite\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-20\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Marcel Hebing <MHebing@diw.de>
Linda Zhu <LZhu@diw.de>

}
\references{
%%~~ Literature or other references for background information ~~
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%~~ simple examples of the most important functions ~~
}
